+++ resources/yin-yang
@@ -1,3 +1,3 @@
 #!/bin/bash
 cd /opt/yin-yang/ || exit
-python3 -O /opt/yin-yang/main.py "$@"
+python3 -Om /opt/yin-yang/yin_yang "$@"
