+++ yin_yang/plugins/kvantum.py
@@ -1,7 +1,7 @@
 import os
 from pathlib import Path
 
-from src.plugins._plugin import PluginCommandline
+from yin_yang.plugins._plugin import PluginCommandline
 
 
 class Kvantum(PluginCommandline):
