+++ yin_yang/plugins/notify.py
@@ -1,4 +1,4 @@
-from src.plugins._plugin import PluginCommandline
+from yin_yang.plugins._plugin import PluginCommandline
 
 
 class Notification(PluginCommandline):
