+++ yin_yang/plugins/system.py
@@ -8,8 +8,8 @@ from pathlib import Path
 
 from PySide6.QtCore import QLocale
 
-from src.meta import Desktop
-from src.plugins._plugin import PluginDesktopDependent, PluginCommandline
+from yin_yang.meta import Desktop
+from yin_yang.plugins._plugin import PluginDesktopDependent, PluginCommandline
 
 logger = logging.getLogger(__name__)
