+++ yin_yang/ui/main_window.py
